chore(p2p_task): rename handler_proptests to handler_proptest
