+++ crates/pathfinder/src/consensus/inner/p2p_task.rs
@@ -53,7 +53,7 @@ use crate::validator::{
 };
 
 #[cfg(test)]
-mod handler_proptests;
+mod handler_proptest;
 #[cfg(test)]
 mod task_tests;
